Added: popup close by Esc
