+++ src/components/constants.js
@@ -1,0 +1,12 @@
+// Constants
+
+export const ESC_KEYCODE = 27;
+
+export const formSettings = {
+  formSelector: '.form',
+  inputSelector: '.form__item',
+  submitButtonSelector: '.form__save-button',
+  inactiveButtonClass: 'form__save-button_disabled',
+  inputErrorClass: 'form__item_type_error',
+  errorClass: 'form__error_visible'
+};
