+++ src/pages/index.js
@@ -2,7 +2,7 @@ import './index.css';
 
 import { enableCards } from '../components/card.js';
 import { enableEditProfile, enableAddCard} from '../components/modal.js';
-import { formSettings } from '../components/utils.js';
+import { formSettings } from '../components/constants.js';
 import { enableValidation} from '../components/validate.js';
 
 enableCards();
